Fixed clear documents method according the new implementation of the file-mediatype index

+++ iserve-sal-core/src/main/java/uk/ac/open/kmi/iserve/sal/manager/impl/DocumentManagerFileSystem.java
@@ -431,8 +431,12 @@ public class DocumentManagerFileSystem extends IntegratedComponent implements Do
      */
     @Override
     public boolean clearDocuments() throws DocumentException {
+
         File mapFile = new File(URI.create(new StringBuilder(documentsInternalPath.toString()).append("/mediaTypeMap.json").toString()));
-        mapFile.delete();
+        if (mapFile.exists()) {
+            mapFile.delete();
+        }
+        fileMediatypeMap.clear();
         File internalFolder = new File(this.getDocumentsInternalPath());
         File[] files = internalFolder.listFiles();
         for (File file : files) {
